pyproject.toml
[tool.poetry]
name = "pubg-scraper"
version = "0.1.0"
description = ""
authors = ["Justin Tervala <jgtervala@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pydantic = "^2.6.4"
requests-ratelimiter = "^0.6.0"
python-dotenv = "^1.0.1"
redis = {extras = ["hiredis"], version = "^5.0.3"}
async-timeout = "^4.0.3"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
black = "^24.2.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
